Audio & Menus Update!
Also worked out a few bugs with the jumpscare scene and got a BGM system integrated in the game base code. Also added some proper graphics to the game. Not reflected here, but I also fleshed out the game page on itch a little more. This game is practically done!